requirements.txt
# Core app & UI
streamlit>=1.36.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Vertex AI (Gemini) SDK
google-genai>=0.3.0

# Auth & Storage
streamlit-authenticator>=0.3.3
SQLAlchemy>=2.0.0
